v0.0.8 - Split Client and API to separate classes for get oportunity receive raw body and mapped DTO
