+++ composer.json
@@ -18,7 +18,7 @@
     }
   ],
   "minimum-stability": "dev",
-  "version": "0.0.7",
+  "version": "0.0.8",
   "autoload": {
     "psr-4": {
       "Shureban\\LaravelSumsubSdk\\": "src/"
